Fix NullReferenceException being thrown when stub method returns null; increase test coverage and split test cases out by category.
